// src/app/interfaces/dash-board.ts
export interface CompraData {
    nombreProveedor: string;
    total: number;
  }
  
  export interface topClientes{
    nombreCliente: string;
    total:number;
  }
  
  export interface topProductos{
    nombreProducto: string;
    totalVendido: number;
  }
  
  export interface utilidad{
    estatus: string;
    total: number;
  }